# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QualiDiagIQ API",
    "description": "\n        Motor de Diagnóstico Tributário Automatizado para a Reforma do Consumo.\n        \n        Parte do ecossistema Tributiq. Fornece análise de maturidade e conformidade\n        com base na EC 132/2023, LC 214/2025 e ABNT NBR 17301:2026.\n        ",
    "contact": {
      "name": "Equipe Tributiq",
      "email": "contato@tributiq.com.br"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/privacidade/solicitacoes": {
      "post": {
        "tags": [
          "Privacidade LGPD"
        ],
        "summary": "Registrar solicitação LGPD do titular",
        "description": "Abre solicitação operacional do art. 18 (acesso, correção, anonimização, eliminação, portabilidade, oposição) no tenant do JWT. Requer header Idempotency-Key.",
        "operationId": "registrar_solicitacao_lgpd_privacidade_solicitacoes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrarSolicitacaoTitularLgpdRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoTitularLgpdResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Privacidade LGPD"
        ],
        "summary": "Listar solicitações LGPD do tenant",
        "description": "Lista solicitações do tenant autenticado, com filtro opcional por status.",
        "operationId": "listar_solicitacoes_lgpd_privacidade_solicitacoes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolicitacaoTitularLgpdResponse"
                  },
                  "title": "Response Listar Solicitacoes Lgpd Privacidade Solicitacoes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/privacidade/solicitacoes/{solicitacao_id}/status": {
      "patch": {
        "tags": [
          "Privacidade LGPD"
        ],
        "summary": "Atualizar status de solicitação LGPD",
        "description": "Atualiza status operacional (backoffice) da solicitação no tenant autenticado.",
        "operationId": "atualizar_status_solicitacao_lgpd_privacidade_solicitacoes__solicitacao_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "solicitacao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Solicitacao Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarStatusSolicitacaoTitularLgpdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoTitularLgpdResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/privacidade/diagnosticos/{diagnostico_id}/anonimizar-respondente": {
      "post": {
        "tags": [
          "Privacidade LGPD"
        ],
        "summary": "Executar anonimização de PII do respondente",
        "description": "Exige solicitação LGPD ``anonimizacao`` com status ``deferida`` ligada ao mesmo ``diagnostico_id``. Regista ``lgpd_anonimizacao_log`` e aplica padrão autorizado pelo WORM (email sentinel, nome marcador, remoção de cargo/telefone).",
        "operationId": "anonimizar_respondente_lgpd_privacidade_diagnosticos__diagnostico_id__anonimizar_respondente_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonimizarRespondenteLgpdHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnonimizarRespondenteLgpdHttpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/privacidade/diagnosticos/{diagnostico_id}/eliminar-diagnostico": {
      "post": {
        "tags": [
          "Privacidade LGPD"
        ],
        "summary": "Eliminar diagnóstico fisicamente (LGPD — pré-finalização)",
        "description": "Exige solicitação LGPD ``eliminacao`` com status ``deferida`` ligada ao mesmo ``diagnostico_id``. Remove a linha apenas se o diagnóstico **não** estiver ``finalizado`` (WORM). Se ``finalizado``, responde **422** orientando anonimização.",
        "operationId": "eliminar_diagnostico_lgpd_privacidade_diagnosticos__diagnostico_id__eliminar_diagnostico_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EliminarDiagnosticoLgpdHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EliminarDiagnosticoLgpdHttpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/privacidade/diagnosticos/{diagnostico_id}/export-portabilidade": {
      "get": {
        "tags": [
          "Privacidade LGPD"
        ],
        "summary": "Exportar pacote de portabilidade (JSON ou PDF com JSON embebido)",
        "description": "Exige solicitação LGPD tipo **portabilidade** com status **deferida** referente ao mesmo `diagnostico_id`. Formato **pacote_pdf**: PDF humano com anexo `qdi-diagnostico-export-v1.json` (ADR-012 §4).",
        "operationId": "export_portabilidade_diagnostico_privacidade_diagnosticos__diagnostico_id__export_portabilidade_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          },
          {
            "name": "solicitacao_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID da solicitação LGPD (tipo portabilidade, status deferida).",
              "title": "Solicitacao Id"
            },
            "description": "UUID da solicitação LGPD (tipo portabilidade, status deferida)."
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FormatoExportPortabilidade",
              "description": "json (default) ou pacote_pdf (PDF com anexo JSON).",
              "default": "json"
            },
            "description": "json (default) ou pacote_pdf (PDF com anexo JSON)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnonimizarRespondenteLgpdHttpRequest": {
        "properties": {
          "solicitacao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Solicitacao Id",
            "description": "Identificador da linha em lgpd_titular_solicitacao (tipo anonimizacao, status deferida)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "solicitacao_id"
        ],
        "title": "AnonimizarRespondenteLgpdHttpRequest",
        "description": "POST /privacidade/diagnosticos/{id}/anonimizar-respondente — exige solicitação ``deferida``."
      },
      "AnonimizarRespondenteLgpdHttpResponse": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnostico Id"
          },
          "solicitacao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Solicitacao Id"
          },
          "status_solicitacao": {
            "type": "string",
            "title": "Status Solicitacao",
            "default": "concluida"
          },
          "mensagem": {
            "type": "string",
            "title": "Mensagem",
            "default": "Anonimização dos campos do respondente aplicada; registo em lgpd_anonimizacao_log (email sentinel + nome marcador + remoção de cargo/telefone)."
          }
        },
        "type": "object",
        "required": [
          "diagnostico_id",
          "solicitacao_id"
        ],
        "title": "AnonimizarRespondenteLgpdHttpResponse",
        "description": "Confirmação após transação (log + UPDATE diagnóstico + conclusão da solicitação)."
      },
      "AtualizarStatusSolicitacaoTitularLgpdRequest": {
        "properties": {
          "status": {
            "type": "string",
            "maxLength": 24,
            "minLength": 4,
            "title": "Status",
            "description": "recebida, em_analise, deferida, indeferida ou concluida."
          },
          "observacao_interna": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacao Interna"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AtualizarStatusSolicitacaoTitularLgpdRequest",
        "description": "PATCH operacional de status da solicitação."
      },
...
      "EliminarDiagnosticoLgpdHttpRequest": {
        "properties": {
          "solicitacao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Solicitacao Id",
            "description": "Identificador da linha em lgpd_titular_solicitacao (tipo eliminacao, status deferida)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "solicitacao_id"
        ],
        "title": "EliminarDiagnosticoLgpdHttpRequest",
        "description": "POST /privacidade/diagnosticos/{id}/eliminar-diagnostico — exige solicitação ``deferida``."
      },
      "EliminarDiagnosticoLgpdHttpResponse": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnostico Id"
          },
          "solicitacao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Solicitacao Id"
          },
          "status_solicitacao": {
            "type": "string",
            "title": "Status Solicitacao",
            "default": "concluida"
          },
          "mensagem": {
            "type": "string",
            "title": "Mensagem",
            "default": "Diagnóstico eliminado fisicamente (estados em_andamento, cancelado ou expirado); solicitação LGPD concluída."
          }
        },
        "type": "object",
        "required": [
          "diagnostico_id",
          "solicitacao_id"
        ],
        "title": "EliminarDiagnosticoLgpdHttpResponse",
        "description": "Confirmação após DELETE do diagnóstico (pré-finalização) e conclusão da solicitação."
      },
...
      "FormatoExportPortabilidade": {
        "type": "string",
        "enum": [
          "json",
          "pacote_pdf"
        ],
        "title": "FormatoExportPortabilidade",
        "description": "Formato do pacote de export (ADR-012 §4)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LgpdTitularSolicitacaoPayload": {
        "properties": {
          "motivo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "detalhes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalhes"
          },
          "idioma_resposta": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Idioma Resposta"
          },
          "referencia_diagnostico_texto": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Referencia Diagnostico Texto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LgpdTitularSolicitacaoPayload",
        "description": "Metadados estruturados da solicitação (art. 18 LGPD) — QDI-H-013."
      },
...
      "RegistrarSolicitacaoTitularLgpdRequest": {
        "properties": {
          "diagnostico_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostico Id",
            "description": "Opcional: vínculo com um diagnóstico específico no tenant."
          },
          "tipo": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "title": "Tipo",
            "description": "acesso, correcao, anonimizacao, eliminacao, portabilidade ou oposicao."
          },
          "canal": {
            "type": "string",
            "maxLength": 24,
            "minLength": 4,
            "title": "Canal",
            "description": "plataforma, self_service ou dpo_email.",
            "default": "plataforma"
          },
          "solicitante_email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5,
            "title": "Solicitante Email"
          },
          "payload": {
            "$ref": "#/components/schemas/LgpdTitularSolicitacaoPayload",
            "description": "Metadados estruturados da solicitação (campos conhecidos; extra proibido)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tipo",
          "solicitante_email"
        ],
        "title": "RegistrarSolicitacaoTitularLgpdRequest",
        "description": "Corpo do POST /privacidade/solicitacoes (art. 18)."
      },
...
      "SolicitacaoTitularLgpdResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "diagnostico_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostico Id"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "canal": {
            "type": "string",
            "title": "Canal"
          },
          "solicitante_email": {
            "type": "string",
            "title": "Solicitante Email"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "observacao_interna": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacao Interna"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "diagnostico_id",
          "tipo",
          "status",
          "canal",
          "solicitante_email",
          "payload",
          "observacao_interna",
          "actor_user_id",
          "criado_em",
          "atualizado_em"
        ],
        "title": "SolicitacaoTitularLgpdResponse",
        "description": "Resposta HTTP da solicitação LGPD registrada."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}